Move the instantiation of GoogleCredential to source. This is needed for GAAP proxy settings.
