+++ gobblin-modules/google-ingestion/src/main/java/gobblin/ingestion/google/webmaster/GoogleWebmasterDataFetcherImpl.java
@@ -22,10 +22,8 @@ import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Queue;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ConcurrentLinkedDeque;
@@ -38,28 +36,19 @@ import java.util.concurrent.TimeoutException;
 
 import org.apache.commons.lang3.tuple.Pair;
 
-import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.googleapis.batch.BatchRequest;
 import com.google.api.client.googleapis.batch.json.JsonBatchCallback;
 import com.google.api.client.repackaged.com.google.common.base.Preconditions;
-import com.google.api.services.webmasters.WebmastersScopes;
 import com.google.api.services.webmasters.model.ApiDimensionFilter;
 import com.google.api.services.webmasters.model.SearchAnalyticsQueryResponse;
 import com.google.common.base.Optional;
 
 import lombok.extern.slf4j.Slf4j;
 
-import gobblin.configuration.ConfigurationKeys;
 import gobblin.configuration.State;
-import gobblin.source.extractor.extract.google.GoogleCommon;
-import gobblin.source.extractor.extract.google.GoogleCommonKeys;
 import gobblin.util.ExecutorsUtils;
 import gobblin.util.limiter.RateBasedLimiter;
 
-import static gobblin.configuration.ConfigurationKeys.SOURCE_CONN_PRIVATE_KEY;
-import static gobblin.configuration.ConfigurationKeys.SOURCE_CONN_USERNAME;
-import static gobblin.configuration.ConfigurationKeys.SOURCE_CONN_USE_PROXY_PORT;
-import static gobblin.configuration.ConfigurationKeys.SOURCE_CONN_USE_PROXY_URL;
 import static gobblin.ingestion.google.webmaster.GoogleWebmasterFilter.Dimension;
 import static gobblin.ingestion.google.webmaster.GoogleWebmasterFilter.FilterOperator;
 import static gobblin.ingestion.google.webmaster.GoogleWebmasterFilter.countryFilterToString;
@@ -76,23 +65,13 @@ public class GoogleWebmasterDataFetcherImpl extends GoogleWebmasterDataFetcher {
   private final GoogleWebmasterClient _client;
   private final List<ProducerJob> _jobs;
 
-  public GoogleWebmasterDataFetcherImpl(State wuState)
-      throws IOException {
-    this(new GoogleWebmasterClientImpl(getCredential(wuState), wuState.getProp(ConfigurationKeys.SOURCE_ENTITY)),
-        wuState);
-  }
-
-  /**
-   * For test only
-   */
   GoogleWebmasterDataFetcherImpl(GoogleWebmasterClient client, State wuState)
       throws IOException {
     _siteProperty = wuState.getProp(GoogleWebMasterSource.KEY_PROPERTY);
     Preconditions.checkArgument(_siteProperty.endsWith("/"), "The site property must end in \"/\"");
     _client = client;
     _jobs = getHotStartJobs(wuState);
-    API_REQUESTS_PER_SECOND =
-        wuState.getPropAsDouble(GoogleWebMasterSource.KEY_PAGES_TUNING_REQUESTS_PER_SECOND, 5.0);
+    API_REQUESTS_PER_SECOND = wuState.getPropAsDouble(GoogleWebMasterSource.KEY_PAGES_TUNING_REQUESTS_PER_SECOND, 5.0);
     GET_PAGE_SIZE_TIME_OUT = wuState.getPropAsInt(GoogleWebMasterSource.KEY_PAGES_TUNING_TIME_OUT, 2);
     LIMITER = new RateBasedLimiter(API_REQUESTS_PER_SECOND, TimeUnit.SECONDS);
     RETRY = wuState.getPropAsInt(GoogleWebMasterSource.KEY_PAGES_TUNING_MAX_RETRIES, 120);
@@ -104,24 +83,6 @@ public class GoogleWebmasterDataFetcherImpl extends GoogleWebmasterDataFetcher {
       return SimpleProducerJob.deserialize(hotStartString);
     }
     return new ArrayList<>();
-  }
-
-  private static Credential getCredential(State wuState) {
-    String scope = wuState.getProp(GoogleCommonKeys.API_SCOPES, WebmastersScopes.WEBMASTERS_READONLY);
-    Preconditions.checkArgument(Objects.equals(WebmastersScopes.WEBMASTERS_READONLY, scope) || Objects
-            .equals(WebmastersScopes.WEBMASTERS, scope),
-        "The scope for WebMaster must either be WEBMASTERS_READONLY or WEBMASTERS");
-
-    String credentialFile = wuState.getProp(SOURCE_CONN_PRIVATE_KEY);
-    List<String> scopes = Collections.singletonList(scope);
-
-//    return GoogleCredential.fromStream(new FileInputStream(credentialFile))
-//        .createScoped(Collections.singletonList(scope));
-
-    return new GoogleCommon.CredentialBuilder(credentialFile, scopes)
-        .fileSystemUri(wuState.getProp(GoogleCommonKeys.PRIVATE_KEY_FILESYSTEM_URI))
-        .proxyUrl(wuState.getProp(SOURCE_CONN_USE_PROXY_URL)).port(wuState.getProp(SOURCE_CONN_USE_PROXY_PORT))
-        .serviceAccountId(wuState.getProp(SOURCE_CONN_USERNAME)).build();
   }
 
   /**
